2022.10.24 Change the flow of the Worker loading CodeIgniter.
